pong-statefun-app/requirements.txt
apache-flink-statefun==3.2.0
aiohttp==3.8.3